# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/client": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get list clients",
                "description": "Returns list clients",
                "operationId": "getClientsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Store new client",
                "description": "Returns client data",
                "operationId": "storeClient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstName": {
                                        "description": "firstName",
                                        "type": "string",
                                        "example": "Piter"
                                    },
                                    "lastName": {
                                        "description": "lastName",
                                        "type": "string",
                                        "example": "Piters"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "123@example.com"
                                    },
                                    "phoneNumber": {
                                        "description": "phoneNumber",
                                        "type": "string",
                                        "example": "+14155552671"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/client/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get client information by id",
                "description": "Returns client data",
                "operationId": "getClientById_",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update existing client",
                "description": "Returns updated project data",
                "operationId": "updateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstName": {
                                        "description": "firstName",
                                        "type": "string",
                                        "example": "Piter"
                                    },
                                    "lastName": {
                                        "description": "lastName",
                                        "type": "string",
                                        "example": "Piters"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "123@example.com"
                                    },
                                    "phoneNumber": {
                                        "description": "phoneNumber",
                                        "type": "string",
                                        "example": "+14155552671"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete existing client",
                "description": "Deletes a record and returns content",
                "operationId": "deleteClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Clients": {
                "title": "Clients",
                "description": "Clients model",
                "properties": {
                    "firstName": {
                        "title": "firstname",
                        "description": "Customer name",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "Customer lastName",
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "Customer email",
                        "type": "string",
                        "format": "string"
                    },
                    "phoneNumber": {
                        "title": "phonenumbar",
                        "description": "Customer phonenumbar",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Clients"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}